Provide the solution for 01-grid-layout-1
I decided to not use fixed tracked sizes for my columns & rows, and
instead utilized fr, because I wanted to have some level of responsive
web design for the exercise, since I could.

I did not explicitly do CSS for the nav nor article elements, because I
did not have to in order to achieve the desired outcome.